--- shell/platform/fuchsia/dart-pkg/zircon/test/meta/channel_test.cml
@@ -1,5 +1,0 @@
-{
-    program: {
-        data: "data/channel_test",
-    },
-}
