+++ shell/platform/fuchsia/dart-pkg/zircon/test/meta/zircon_tests.cml
@@ -1,0 +1,6 @@
+{
+    program: {
+        data: "data/zircon_tests",
+        runner: "dart_jit_runner",
+    },
+}
