+++ shell/platform/fuchsia/dart/BUILD.gn
@@ -96,6 +96,7 @@ dart_library("expect") {
   deps = [ ":meta" ]
 
   sources = [
+    "config.dart",
     "expect.dart",
     "minitest.dart",
   ]
