+++ shell/platform/fuchsia/dart_runner/dart_component_controller.h
@@ -5,18 +5,22 @@
 #ifndef FLUTTER_SHELL_PLATFORM_FUCHSIA_DART_RUNNER_DART_COMPONENT_CONTROLLER_H_
 #define FLUTTER_SHELL_PLATFORM_FUCHSIA_DART_RUNNER_DART_COMPONENT_CONTROLLER_H_
 
-#include <memory>
-
+#include <dart/test/cpp/fidl.h>
 #include <fuchsia/component/runner/cpp/fidl.h>
 #include <fuchsia/sys/cpp/fidl.h>
 #include <lib/async-loop/cpp/loop.h>
 #include <lib/async/cpp/wait.h>
 #include <lib/fdio/namespace.h>
+#include <lib/fidl/cpp/binding.h>
+#include <lib/fidl/cpp/binding_set.h>
+#include <lib/fidl/cpp/string.h>
 #include <lib/sys/cpp/component_context.h>
 #include <lib/sys/cpp/service_directory.h>
+#include <lib/vfs/cpp/pseudo_dir.h>
 #include <lib/zx/timer.h>
 
-#include "lib/fidl/cpp/binding.h"
+#include <memory>
+
 #include "runtime/dart/utils/mapped_resource.h"
 #include "third_party/dart/runtime/include/dart_api.h"
 
@@ -24,7 +28,8 @@ namespace dart_runner {
 
 /// Starts a Dart component written in CFv2.
 class DartComponentController
-    : public fuchsia::component::runner::ComponentController {
+    : public dart::test::Echo,
+      public fuchsia::component::runner::ComponentController {
  public:
   DartComponentController(
       fuchsia::component::runner::ComponentStartInfo start_info,
@@ -59,6 +64,9 @@ class DartComponentController
   bool CreateIsolate(const uint8_t* isolate_snapshot_data,
                      const uint8_t* isolate_snapshot_instructions);
 
+  // |Echo|
+  void EchoString(fidl::StringPtr value, EchoStringCallback callback) override;
+
   // |ComponentController|
   void Kill() override;
   void Stop() override;
@@ -76,12 +84,17 @@ class DartComponentController
 
   std::string label_;
   std::string url_;
-  std::shared_ptr<sys::ServiceDirectory> runner_incoming_services_;
   std::string data_path_;
+  std::shared_ptr<sys::ServiceDirectory> runner_incoming_services_;
   std::unique_ptr<sys::ComponentContext> context_;
+  std::unique_ptr<vfs::PseudoDir> dart_outgoing_dir_;
+  fuchsia::io::DirectoryPtr dart_outgoing_dir_ptr_;
+  fidl::InterfaceRequest<fuchsia::io::Directory> dart_outgoing_dir_request_;
+  fuchsia::io::NodePtr dart_outgoing_dir_ptr_to_check_on_open_;
 
   fuchsia::component::runner::ComponentStartInfo start_info_;
   fidl::Binding<fuchsia::component::runner::ComponentController> binding_;
+  fidl::BindingSet<dart::test::Echo> echo_binding_;
 
   fdio_ns_t* namespace_ = nullptr;
   int stdout_fd_ = -1;
