+++ shell/platform/fuchsia/dart_runner/fidl/echo.fidl
@@ -1,8 +1,8 @@
-// Copyright 2019 The Fuchsia Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-library flutter.example.echo;
+library dart.test;
 
 const MAX_STRING_LENGTH uint64 = 32;
