+++ shell/platform/fuchsia/dart_runner/fidl/meta.json
@@ -1,0 +1,9 @@
+{
+    "deps": [],
+    "name": "dart.test",
+    "root": "dart.test",
+    "sources": [
+        "../../../flutter/shell/platform/fuchsia/dart_runner/fidl/echo.fidl"
+    ],
+    "type": "fidl_library"
+}
