--- shell/platform/fuchsia/dart_runner/tests/fidl/flutter.example.echo/README.md
@@ -1,3 +1,0 @@
-# Echo FIDL
-
-This FIDL protocol allows the Dart integration tests to communicate with a Dart Echo server (a test Dart component that echoes back a request).
