+++ shell/platform/fuchsia/flutter/tests/integration/embedder/parent-view/BUILD.gn
@@ -14,8 +14,6 @@ dart_library("lib") {
   deps = [
     "//flutter/shell/platform/fuchsia/dart:args",
     "//flutter/shell/platform/fuchsia/dart:vector_math",
-    "//flutter/tools/fuchsia/dart:fuchsia_services",
-    "//flutter/tools/fuchsia/dart:zircon",
     "//flutter/tools/fuchsia/fidl:fuchsia.sys",
     "//flutter/tools/fuchsia/fidl:fuchsia.ui.app",
     "//flutter/tools/fuchsia/fidl:fuchsia.ui.views",
