+++ shell/platform/fuchsia/flutter/tests/integration/touch-input/embedding-flutter-view/lib/embedding-flutter-view.dart
@@ -7,10 +7,10 @@ import 'dart:convert';
 import 'dart:typed_data';
 import 'dart:io';
 import 'dart:ui';
+import 'dart:zircon';
 
 import 'package:args/args.dart';
 import 'package:vector_math/vector_math_64.dart' as vector_math_64;
-import 'package:zircon/zircon.dart';
 
 final _argsCsvFilePath = '/config/data/args.csv';
