+++ shell/platform/fuchsia/flutter/tests/integration/touch-input/touch-input-view/lib/touch-input-view.dart
@@ -5,8 +5,7 @@
 import 'dart:convert';
 import 'dart:typed_data';
 import 'dart:ui';
-
-import 'package:zircon/zircon.dart';
+import 'dart:zircon';
 
 void main() {
   print('Launching touch-input-view');
